# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ItineraryInputBudgetTier = typeof ItineraryInputBudgetTier[keyof typeof ItineraryInputBudgetTier];


export const ItineraryInputBudgetTier = {
  Premium: 'Premium',
  'Ultra-Luxury': 'Ultra-Luxury',
} as const;

export type ItineraryInputSpecialOccasion = typeof ItineraryInputSpecialOccasion[keyof typeof ItineraryInputSpecialOccasion];


export const ItineraryInputSpecialOccasion = {
  None: 'None',
  Anniversary: 'Anniversary',
  Honeymoon: 'Honeymoon',
  Birthday: 'Birthday',
  Family_Reunion: 'Family Reunion',
} as const;

export interface ItineraryInput {
  guestName: string;
  checkIn: string;
  checkOut: string;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children: number;
  interests: string[];
  budgetTier: ItineraryInputBudgetTier;
  specialOccasion: ItineraryInputSpecialOccasion;
  /** @nullable */
  specialNotes?: string | null;
}

export interface ItineraryActivity {
  time: string;
  name: string;
  description: string;
  duration: string;
  pricePerPerson: number;
  unsplashKeyword: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface ItineraryDay {
  day: number;
  date: string;
  activities: ItineraryActivity[];
}

export interface Itinerary {
  id: string;
  guestName: string;
  checkIn: string;
  checkOut: string;
  adults: number;
  children: number;
  interests: string[];
  budgetTier: string;
  specialOccasion: string;
  /** @nullable */
  specialNotes?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  days: ItineraryDay[];
  approved: boolean;
  createdAt: string;
}

export interface Activity {
  name: string;
  category: string;
  description: string;
  duration: string;
  pricePerPerson: number;
  bestFor: string;
  timeOfDay: string;
  unsplashKeyword: string;
}