# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pure Kauai Luxury Travel Concierge API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: itineraries
    description: Itinerary operations
  - name: activities
    description: Activity catalog operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /itineraries:
    post:
      operationId: createItinerary
      tags: [itineraries]
      summary: Generate a new itinerary using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItineraryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /itineraries/{id}:
    get:
      operationId: getItinerary
      tags: [itineraries]
      summary: Get an itinerary by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /itineraries/{id}/approve:
    post:
      operationId: approveItinerary
      tags: [itineraries]
      summary: Approve an itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all available activities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ItineraryInput:
      type: object
      required:
        - guestName
        - checkIn
        - checkOut
        - adults
        - children
        - interests
        - budgetTier
        - specialOccasion
      properties:
        guestName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        adults:
          type: integer
          minimum: 1
        children:
          type: integer
          minimum: 0
        interests:
          type: array
          items:
            type: string
        budgetTier:
          type: string
          enum: [Premium, Ultra-Luxury]
        specialOccasion:
          type: string
          enum: [None, Anniversary, Honeymoon, Birthday, Family Reunion]
        specialNotes:
          type: ["string", "null"]

    ItineraryDay:
      type: object
      required:
        - day
        - date
        - activities
      properties:
        day:
          type: integer
        date:
          type: string
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryActivity"

    ItineraryActivity:
      type: object
      required:
        - time
        - name
        - description
        - duration
        - pricePerPerson
        - unsplashKeyword
      properties:
        time:
          type: string
        name:
          type: string
        description:
          type: string
        duration:
          type: string
        pricePerPerson:
          type: number
        unsplashKeyword:
          type: string
        photoUrl:
          type: ["string", "null"]

    Itinerary:
      type: object
      required:
        - id
        - guestName
        - checkIn
        - checkOut
        - adults
        - children
        - interests
        - budgetTier
        - specialOccasion
        - days
        - approved
        - createdAt
      properties:
        id:
          type: string
        guestName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        adults:
          type: integer
        children:
          type: integer
        interests:
          type: array
          items:
            type: string
        budgetTier:
          type: string
        specialOccasion:
          type: string
        specialNotes:
          type: ["string", "null"]
        welcomeMessage:
          type: ["string", "null"]
        days:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDay"
        approved:
          type: boolean
        createdAt:
          type: string

    Activity:
      type: object
      required:
        - name
        - category
        - description
        - duration
        - pricePerPerson
        - bestFor
        - timeOfDay
        - unsplashKeyword
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        duration:
          type: string
        pricePerPerson:
          type: number
        bestFor:
          type: string
        timeOfDay:
          type: string
        unsplashKeyword:
          type: string