# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  name: string;
  category: string;
  description: string;
  duration: string;
  pricePerPerson: number;
  bestFor: string;
  timeOfDay: string;
  unsplashKeyword: string;
}