# lib/api-zod/src/generated/types/itinerary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryDay } from './itineraryDay';

export interface Itinerary {
  id: string;
  guestName: string;
  checkIn: string;
  checkOut: string;
  adults: number;
  children: number;
  interests: string[];
  budgetTier: string;
  specialOccasion: string;
  /** @nullable */
  specialNotes?: string | null;
  /** @nullable */
  welcomeMessage?: string | null;
  days: ItineraryDay[];
  approved: boolean;
  createdAt: string;
}