# lib/api-zod/src/generated/types/itineraryActivity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryActivity {
  time: string;
  name: string;
  description: string;
  duration: string;
  pricePerPerson: number;
  unsplashKeyword: string;
  /** @nullable */
  photoUrl?: string | null;
}