# lib/api-zod/src/generated/types/itineraryDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryActivity } from './itineraryActivity';

export interface ItineraryDay {
  day: number;
  date: string;
  activities: ItineraryActivity[];
}