# lib/api-zod/src/generated/types/itineraryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pure Kauai Luxury Travel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryInputBudgetTier } from './itineraryInputBudgetTier';
import type { ItineraryInputSpecialOccasion } from './itineraryInputSpecialOccasion';

export interface ItineraryInput {
  guestName: string;
  checkIn: string;
  checkOut: string;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children: number;
  interests: string[];
  budgetTier: ItineraryInputBudgetTier;
  specialOccasion: ItineraryInputSpecialOccasion;
  /** @nullable */
  specialNotes?: string | null;
}